Release 0.4.2 — float more number-typed readings (#352 follow-up)
Mirror the Twinguard temperature fix for the other readings the local API
types as number but the lib truncated with int(): AirQualityLevelService
humidity + purity (Twinguard) and ValveTappetService position (Thermostat
II). All now return float; matching model properties int -> float. Restores
decimal resolution in the recorder/statistics. Adds regression tests.

+++ boschshcpy/models_impl.py
@@ -1543,7 +1543,7 @@ class SHCThermostat(
         )  # type: ignore[assignment]
 
     @property
-    def position(self) -> int:
+    def position(self) -> float:
         return self._valvetappet_service.position
 
     @property
@@ -2524,7 +2524,7 @@ class SHCTwinguard(SHCBatteryDevice):
         return self._airqualitylevel_service.temperatureRating
 
     @property
-    def humidity(self) -> int:
+    def humidity(self) -> float:
         return self._airqualitylevel_service.humidity
 
     @property
@@ -2532,7 +2532,7 @@ class SHCTwinguard(SHCBatteryDevice):
         return self._airqualitylevel_service.humidityRating
 
     @property
-    def purity(self) -> int:
+    def purity(self) -> float:
         return self._airqualitylevel_service.purity
 
     @property
